rename components/sosContacts to components/sosContactList, add tests on soscontactlist
