+++ src/components/sosContactList/index.js
@@ -53,7 +53,9 @@ export default function SosContactList({ contacts, navigation }) {
         />
       </View>
       {contacts !== undefined && contacts.length > 0 && (
-        <StyledButton style={styles.messageButtonContainer}>
+        <StyledButton
+          style={styles.messageButtonContainer}
+          testID='send-sms-button'>
           {/* eslint-disable-next-line react-native/no-raw-text */}
           <StyledButtonText style={styles.messageButtonText}>
             Ask for help to your contacts
